// core/types/editor.ts
export enum ComponentGroup {
    UI,
    Content,
    Grid,
    Utils,
}
// ...
export type FormFieldConfig =
    | FormFieldConfigPrototype
    | FormBasicFieldConfig
    | FormNumberFieldConfig
    | FormArrayFieldConfig
    | FormComplexFieldConfig
    | FormStringFieldConfig
    | FormFileFieldConfig
    | FormPasswordFieldConfig
    | FormTextFieldConfig
    | FormWysiwygFieldConfig
    | FormGalleryFieldConfig
    | FormImageFieldConfig
    | FormColorFieldConfig
    | FormSelectFieldConfig;

/**
 * Options for component selector component
 */
export type ComponentSelectorOption = {
    value: string;
    text: string;
};

/**
 * Block description type
 */
export type ComponentBlockType = {
    name: string;
    options: { [key: string]: any };
    editableOptions?: FormConfig;
};

/**
 * Editable component options type
 */
export type FormConfig = {
    [name: string]: FormFieldConfig;
};
// ...
export type ComponentConfig = {
    /** Component name */
    name: string;
    /** Options configuration for editor */
    options: FormConfig;
    /** Whether or not component supports nesting */
    has_children?: boolean;
    /**
     * If component supports nesting can be set to prevent delete operation
     * from removing a child from the children list and instead leave the spot empty
     * so that another component can take its place. This keeps the children
     * list count and positions/indexes the same.
     */
    enforce_child_count?: boolean;
    /** Component name or short description in component selector */
    selector_text: string;
    /** Full component description for editor */
    editor_description: string;
    /** Component group */
    group: ComponentGroup;
};